+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Fixed
+
+- Fix invalid refspec when deleting the ticket branch
+
 ## 4.1.0 - 2026-08-14
 
 ### Changed